requirement.txt
accelerate==1.12.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
dlt==1.20.0
docling==2.66.0
docling-core==2.57.0
docling-ibm-models==3.10.3
docling-parse==4.7.2
et-xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
faker==40.1.0
fastapi==0.128.0
fastjsonschema==2.21.2
filelock==3.20.2
filetype==1.2.0
fonttools==4.61.1
fqdn==1.5.1
fsspec==2025.12.0
gitdb==4.0.12
gitpython==3.1.46
giturlparse==0.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanize==4.15.0
idna==3.11
ipykernel==7.1.0
ipython==8.38.0
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.12.0
json5==0.13.0
jsonlines==4.0.0
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.7.0
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.5.1
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
kiwisolver==1.4.9
lance-namespace==0.4.3
lance-namespace-urllib3-client==0.4.3
lancedb==0.26.1
lark==1.3.1
latex2mathml==3.78.1
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.2
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook-shim==0.2.4
numpy==2.2.6
omegaconf==2.3.0
openai==2.14.0
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathvalidate==3.3.1
pendulum==3.1.0
pillow==11.3.0
platformdirs==4.5.1
pluggy==1.6.0
ply==3.11
polyfactory==3.2.0
prometheus-client==0.23.1
prompt-toolkit==3.0.52
psutil==7.2.1
pure-eval==0.2.3
pyclipper==1.4.0
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pylatexenc==2.10
pyparsing==3.3.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
python-pptx==1.0.2
pytz==2025.2
pywinpty==3.0.2
pyyaml==6.0.3
pyzmq==27.1.0
rapidocr==3.4.5
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requirements-parser==0.13.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-argparse==1.7.2
rpds-py==0.30.0
rtree==1.4.1
safetensors==0.7.0
scipy==1.15.3
seaborn==0.13.2
semchunk==2.2.2
semver==3.0.4
send2trash==2.0.0
shapely==2.1.2
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
sqlglot==28.5.0
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.2
uvicorn==0.40.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
win-precise-time==1.4.2
xlsxwriter==3.2.9